Move MemberNFT interface to nfts/interfaces. Organize.
