+++ packages/webapp/src/members/components/member-chip-components/nft-info.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { atomicAssets } from "config";
 import { assetToLocaleString, openInNewTab, useCountdown } from "_app";
 import { NFT } from "_app/ui/icons";
-import { AssetData, MemberAuctionData, MemberNFT } from "members";
+import { MemberNFTAssetData, MemberNFTAuctionData, MemberNFT } from "nfts";
 
 export const NFTInfo = ({ member }: { member: MemberNFT }) => {
     if (member.auctionData) {
@@ -23,7 +23,11 @@ export const NFTInfo = ({ member }: { member: MemberNFT }) => {
 
 export default NFTInfo;
 
-const AuctionBadge = ({ auctionData }: { auctionData: MemberAuctionData }) => {
+const AuctionBadge = ({
+    auctionData,
+}: {
+    auctionData: MemberNFTAuctionData;
+}) => {
     const countdown = useCountdown({
         endTime: new Date(auctionData.bidEndTime as number),
         interval: 30000,
@@ -52,7 +56,7 @@ const SaleBadge = ({
     assetData,
     saleId,
 }: {
-    assetData: AssetData;
+    assetData: MemberNFTAssetData;
     saleId: string;
 }) => (
     <div
@@ -68,7 +72,7 @@ const SaleBadge = ({
     </div>
 );
 
-const AssetBadge = ({ assetData }: { assetData: AssetData }) => (
+const AssetBadge = ({ assetData }: { assetData: MemberNFTAssetData }) => (
     <NFTBadge
         onClick={(e) => {
             e.stopPropagation();
